+++ relay-event-schema/src/protocol/span/convert.rs
@@ -17,7 +17,6 @@ impl From<&Event> for Span {
 
             measurements,
             _metrics,
-            performance_issues_spans,
             ..
         } = event;
 
@@ -81,7 +80,6 @@ impl From<&Event> for Span {
             platform: platform.clone(),
             was_transaction: true.into(),
             kind: Default::default(),
-            performance_issues_spans: performance_issues_spans.clone(),
             other: Default::default(),
         }
     }
@@ -148,7 +146,7 @@ mod tests {
         .unwrap();
 
         let span_from_event = Span::from(&event);
-        insta::assert_debug_snapshot!(span_from_event, @r###"
+        insta::assert_debug_snapshot!(span_from_event, @r#"
         Span {
             timestamp: ~,
             start_timestamp: ~,
@@ -260,6 +258,9 @@ mod tests {
                 lcp_size: ~,
                 lcp_id: ~,
                 lcp_url: ~,
+                sentry_dsc_trace_id: ~,
+                sentry_dsc_transaction: ~,
+                sentry_dsc_project_id: ~,
                 span_name: "my 1st transaction",
                 other: {
                     "custom_attribute": I64(
@@ -295,9 +296,8 @@ mod tests {
             platform: "php",
             was_transaction: true,
             kind: ~,
-            performance_issues_spans: ~,
             other: {},
         }
-        "###);
+        "#);
     }
 }