+++ tests/integration/test_nel.py
@@ -66,6 +66,10 @@ def test_nel_converted_to_logs(mini_sentry, relay):
                         "value": "Firefox",
                     },
                     "browser.version": {"type": "string", "value": "42.0"},
+                    "user_agent.original": {
+                        "type": "string",
+                        "value": "RelayIntegrationTests/1.0.0 Firefox/42.0",
+                    },
                     "sentry.observed_timestamp_nanos": {
                         "type": "string",
                         "value": time_within_delta(expect_resolution="ns"),